Bug fixin
* Changed out Int to BigInteger
* Added distinct to Tx query so it actually only brings back tx objects/counts
* Added total tx count to Spotlight object
* Made Block Hash look like a hash
* Added the protos that should be there
* Updated all the proto-related lists
* Added basic NFT count to AssetDetail
* Updated a couple docker scripts cuz I moved things around in the last commit
* Because we have a docker gha workflow that builds, I removed the branch-> main designation for build gha

closes: #34

+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/CommonModels.kt
@@ -25,8 +25,8 @@ data class CoinDec (val amount: BigDecimal, val denom: String)
 fun CoinOuterClass.Coin.toData() = Coin(this.amount.toBigInteger(), this.denom)
 
 data class CountTotal(
-    val count: Int,
-    val total: Int
+    val count: BigInteger,
+    val total: BigInteger
 )
 
 data class BondedTokens(
